hamgit-projects/pdf-generator/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Brotli==1.1.0
cffi==1.17.0
click==8.1.7
cssselect2==0.7.0
fastapi==0.112.2
fonttools==4.53.1
h11==0.14.0
html5lib==1.1
httptools==0.6.1
idna==3.8
jalali_core==1.0.0
jdatetime==5.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydyf==0.11.0
pyphen==0.16.0
python-dotenv==1.0.1
PyYAML==6.0.2
qrcode==8.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tinycss2==1.3.0
typing_extensions==4.12.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.24.0
weasyprint==62.3
webencodings==0.5.1
websockets==13.0.1
zopfli==0.2.3
